Conflict p-values

A test of the hypothesis that the direct data on a disease outcome give the same information about that outcome as an indirect evidence synthesis obtained from a fitted disbayes model. The outcome may be annual incidence, mortality, remission probabilities, or prevalence.

conflict_disbayes(x, varname)

Arguments

x

A fitted disbayes model.

varname

Either inc, prev, mort or rem.

Value

A data frame with columns indicating age, gender and area.

p1 is a "one-sided" p-value for the null hypothesis that \(r_{obs}=r_{fit}\) against the alternative that \(r_{obs} > r_{fit}\),

p2 is the two-sided p-value for the null hypothesis that \(r_{obs}=r_{fit}\) against the alternative that \(r_{obs}\) is not equal to \(r_{fit}\),

where \(r_{obs}\) is the rate informed only by direct data, and \(r_{fit}\) is the rate informed by evidence synthesis. Therefore if the evidence synthesis excludes the direct data, then these are interpreted as "conflict" p-values (see Presanis et al. 2013).

In each case, a small p-value favours the alternative hypothesis.

Details

Hierarchical models are not currently supported in this function.

References

Presanis, A. M., Ohlssen, D., Spiegelhalter, D. J. and De Angelis, D. (2013) Conflict diagnostics in directed acyclic graphs, with applications in Bayesian evidence synthesis. Statistical Science, 28, 376-397.