Bayesian estimation of chronic disease epidemiology from incomplete data

Estimates a three-state disease model given data on incidence, prevalence and mortality.

disbayes(
  data,
  inc_num = NULL,
  inc_denom = NULL,
  inc = NULL,
  inc_lower = NULL,
  inc_upper = NULL,
  prev_num = NULL,
  prev_denom = NULL,
  prev = NULL,
  prev_lower = NULL,
  prev_upper = NULL,
  mort_num = NULL,
  mort_denom = NULL,
  mort = NULL,
  mort_lower = NULL,
  mort_upper = NULL,
  rem_num = NULL,
  rem_denom = NULL,
  rem = NULL,
  rem_lower = NULL,
  rem_upper = NULL,
  age = "age",
  cf_model = "smooth",
  inc_model = "smooth",
  rem_model = "const",
  prev_zero = FALSE,
  inc_trend = NULL,
  cf_trend = NULL,
  cf_init = 0.01,
  eqage = 30,
  eqagehi = NULL,
  loo = TRUE,
  sprior = c(1, 1),
  scf_fixed = NULL,
  sinc_fixed = NULL,
  rem_prior = c(1.1, 1),
  inc_prior = c(2, 0.1),
  cf_prior = c(2, 0.1),
  method = "opt",
  draws = 1000,
  iter = 10000,
  stan_control = NULL,
  bias_model = NULL,
  ...
)

Arguments

data
Data frame containing some of the variables below. The variables below are provided as character strings naming columns in this data frame. For each disease measure available, one of the following three combinations of variables must be specified:

(1) numerator and denominator (2) estimate and denominator (3) estimate with lower and upper credible limits

Mortality must be supplied, and at least one of incidence and prevalence. If remission is assumed to be possible, then remission data should also be supplied (see below).

For estimates based on registry data assumed to cover the whole population, then the denominator will be the population size.

inc_num
Numerator for the incidence data, assumed to represent the observed number of new cases within a year among a population of size inc_denom.

inc_denom
Denominator for the incidence data.

The function ci2num can be used to convert a published estimate and interval for a proportion to an implicit numerator and denominator.

Note that to include extra uncertainty beyond that implied by a published interval, the numerator and denominator could be multiplied by a constant, for example, multiplying both the numerator and denominator by 0.5 would give the data source half its original weight.

inc
Estimate of incidence.

inc_lower
Lower credible limit for the incidence estimate

inc_upper
Upper credible limit for the incidence estimate

prev_num
Numerator for the estimate of prevalence

prev_denom
Denominator for the estimate of prevalence (e.g. the size of the survey used to obtain the prevalence estimate)

prev
Estimate of prevalence

prev_lower
Lower credible limit for the prevalence estimate

prev_upper
Upper credible limit for the prevalence estimate

mort_num
Numerator for the estimate of the mortality rate

mort_denom
Denominator for the estimate of the mortality rate (e.g. the population size, if the estimates were obtained from a comprehensive register)

mort
Estimate of the mortality rate

mort_lower
Lower credible limit for the mortality estimate

mort_upper
Upper credible limit for the mortality estimate

rem_num
Numerator for the estimate of the remission rate. Remission data should be supplied if remission is permitted in the model, either as a numerator and denominator or as an estimate and lower credible interval. Conversely, if no remission data are supplied, then remission is assumed to be impossible. These "data" may represent a prior judgement rather than observation - lower denominators or wider credible limits represent weaker prior information.

rem_denom
Denominator for the estimate of the remission rate

rem
Estimate of the remission rate

rem_lower
Lower credible limit for the remission estimate

rem_upper
Upper credible limit for the remission estimate

age
Variable in the data indicating the year of age

cf_model
Model for how case fatality varies with age.

"smooth" (the default). Case fatality is modelled as a smooth function of age, using a spline.

"indep" Case fatalities are estimated independently for each year of age. This may be useful for determining how much information is in the data. That is, if the posterior from this model is identical to the prior for a certain age, then there is no information in the data alone about case fatality at that age, indicating that some other structural assumption (such as a smooth function of age) or external data are equired to give more precise estimates.

"increasing" Case fatality is modelled as a smooth and increasing function of age.

"const" Case fatality is modelled as constant with age.

inc_model
Model for how incidence varies with age.

"smooth" (the default). Incidence is modelled as a smooth spline function of age.

"indep" Incidence rates for each year of age are estimated independently.

rem_model
Model for how remission rates vary with age, which are typically less well-informed by data, compared to incidence and case fatality.

"const" (the default). Constant remission rate over all ages.

"indep" Remission rates estimated independently over all ages.

prev_zero
If TRUE, attempt to estimate prevalence at age zero from the data, as part of the Bayesian model, even if the observed prevalence is zero. Otherwise (the default) this is assumed to be zero if the count is zero, and estimated otherwise.

inc_trend
Matrix of constants representing trends in incidence through calendar time by year of age. There are nage rows and nage columns, where nage is the number of years of age represented in the data. The entry in the ith row and jth column represents the ratio between the incidence nage+j years prior to the year of the data, year, and the incidence in the year of the data, for a person i years of age. For example, if nage=100 and the data refer to the year 2017, then the first column refers to the year 1918 and the last (100th) column refers to 2017. The last column should be all 1, unless the current data are supposed to be biased.

To produce this format from a long data frame, filter to the appropriate outcome and subgroup, and use pivot_wider, e.g.

trends <- ihdtrends filter(outcome=="Incidence", gender=="Female") pivot_wider(names_from="year", values_from="p2017") select(-age, -gender, -outcome) as.matrix()

cf_trend
Matrix of constants representing trends in case fatality through calendar time by year of age, in the same format as inc_trend.

cf_init
Initial guess at a typical case fatality value, for an average age.

eqage
Case fatalities (and incidences) are assumed to be equal for all ages below this age, inclusive, when using the smoothed model.

eqagehi
Case fatalities (and incidences) are assumed to be equal for all ages above this age, inclusive, when using the smoothed model.

loo
Compute leave-one-out cross validation statistics (for method="mcmc" only).

sprior
Vector of two elements, giving the rates of the exponential prior distributions used to penalise the coefficients of the spline model. The first refers to the spline model for incidence, the second for case fatality. The default of 1 should adapt appropriately to the data, but higher values give stronger smoothing, or lower values give weaker smoothing, if required.

scf_fixed
If this is set to a number, then the parameter determining the smoothness of case fatality with age (i.e. the standard deviation of the spline basis coefficients) is fixed at this number. If scf_fixed=TRUE, then an "empirical Bayes" method is used, where the posterior mode of this parameter is determined (using the prior defined by sprior), and the model is fitted again with the parameter fixed at this mode. If scf_fixed=NULL (the default) then a "full Bayes" method is used where the full posterior of this parameter is estimated. NOT IMPLEMENTED YET FOR HIERARCHICAL MODELS

sinc_fixed
As scf_fixed, but for incidence instead of case fatality.

rem_prior
Vector of two elements giving the Gamma shape and rate parameters of the prior for the remission rate, used in both rem_model="const" and rem_model="fixed".

inc_prior
Vector of two elements giving the Gamma shape and rate parameters of the prior for the incidence rate. Only used if inc_model="indep", for each age-specific rate.

cf_prior
Vector of two elements giving the Gamma shape and rate parameters of the prior for the case fatality rate. Only used if cf_model="indep", for each age-specific rate, and for the rate at eqage in cf_model="increasing".

method
String indicating the inference method, defaulting to "opt".

If method="mcmc" then a sample from the posterior is drawn using Markov Chain Monte Carlo sampling, via rstan's rstan::sampling() function. This is the most accurate but the slowest method.

If method="opt", then instead of an MCMC sample from the posterior, `disbayes` returns the posterior mode calculated using optimisation, via rstan's rstan::optimizing() function. A sample from a normal approximation to the (real-line-transformed) posterior distribution is drawn in order to obtain credible intervals.

If the optimisation fails to converge (non-zero return code), try increasing the number of iterations from the default 1000, e.g. `disbayes(..., iter=10000, ...)`, or changing the algorithm to `disbayes(..., algorithm="Newton", ...)`.

If there is an error message that mentions `chol`, then the computed Hessian matrix is not positive definite at the reported optimum, hence credible intervals cannot be computed. This can occur either because of numerical error in computation of the Hessian, or because the reported optimum is wrong. If you are willing to believe the optimum and live without credible intervals, then set draws=0 to skip computation of the Hessian. To examine the problematic Hessian, set hessian=TRUE,draws=0, then look at the $fit$hessian component of the `disbayes` return object. If it can be inverted, do sqrt(diag(solve())) on the Hessian, and check for NaNs, indicating the problematic parameters. Otherwise, diagonal entries of the Hessian matrix that are very small may indicate parameters that are poorly identified from the data, leading to computational problems.

If method="vb", then variational Bayes methods are used, via rstan's rstan::vb() function. This is labelled as "experimental" by rstan. It might give a better approximation to the posterior than method="opt", but has not been investigated much for `disbayes` models.

draws
Number of draws from the normal approximation to the posterior when using method="opt".

iter
Number of iterations for MCMC sampling, or maximum number of iterations for optimization.

stan_control
(method="mcmc" only). List of options supplied as the control argument to rstan::sampling() in rstan for the main model fit.

bias_model
Experimental model for bias in the incidence estimates due to conflicting information between the different data sources. If bias_model=NULL (the default) no bias is assumed, and all data are assumed to be generated from the same age-specific incidences.

Otherwise there are assumed to be two alternative curves of incidence by age (denoted 2 and 1) where curve 2 is related to curve 1 via a constant hazard ratio that is estimated from the data, given a standard normal prior on the log scale. Three distinct curves would not be identifiable from the data.

If bias_model="inc" then the incidence data is assumed to be generated from curve 2, and the prevalence and mortality data from curve 1.

bias_model="prev" then the prevalence data is generated from curve 2, and the incidence and mortality data from curve 1.

If bias_model="incprev" then both incidence and prevalence data are generated from curve 2, and the mortality data from curve 1.

...
Further arguments passed to rstan::sampling() to control MCMC sampling, or rstan::optimizing() to control optimisation, in Stan.

Value

A list with the following components

fit: An object containing posterior samples from the fitted model, in the stanfit format returned by the stan function in the rstan package.

loo: A list of objects containing leave-one-out cross-validation statistics. There is one list component for each of the observed outcomes informing the model: incidence, prevalence, mortality and remission. The component for each outcome is an object in the form returned by the loo function in the loo package. This can be used to assess how well the model predicts the data for that outcome, compared to other models. The looi_disbayes function can be used to extract from this list a single tidy data frame with one row per observation.

dat: A list containing the input data in the form of numerators and denominators.

Use the tidy.disbayes method to return summary statistics from the fitted models, simply by calling tidy() on the fitted model.

Details

It is designed to estimate case fatality and incidence, given data on mortality and at least one of incidence and prevalence. Remission may also be included.