Trends in ischemic heart disease in England

Trends in ischemic heart disease in England

ihdtrends

Format

A data frame with columns:

gender: "male" or "female".

age: Year of age.

year: Calendar year.

p2017: Estimated ratio between the outcome in the calendar year and the outcome in 2017.

outcome: Outcome referred to (incidence or case fatality).

Source

Scarborough, P., Wickramasinghe, K., Bhatnagar, P. and Rayner, M. (2011) Trends in coronary heart disease, 1961-2001. British Heart Foundation.

Smolina, K., Wright, F. L., Rayner, M. and Goldacre, M. J. (2012) Determinants of the decline in mortality from acute myocardial infarction in England between 2002 and 2010: linked national database study. BMJ, 344.

British Heart Foundation (2020) Heart and Circulatory Disease Statistics 2020. British Heart Foundation.

Details

The data were interpolated and smoothed to produce a matrix by year of age and calendar year, using the script at https://github.com/chjackson/disbayes/blob/master/data-raw/trends.r.