Extract observation-specific contributions from a disbayes leave-one-out cross validation

Extract observation-specific contributions from a disbayes leave-one-out cross validation

loo_indiv(x, agg = FALSE)

looi_disbayes(x, agg = FALSE)

Arguments

x
For loo_indiv, an object returned by loo.disbayes. For looi_disbayes, an object returned by disbayes.

agg
If TRUE then the observation-specific contributions are aggregated over outcome type, returning a data frame with one row for each of incidence, prevalence, mortality and remission (if remission is included in the model), and one column for each of "elpd_loo", "p_loo" and "looic".

Value

A data frame with one row per observed age-specific mortality, incidence, prevalence and/or remission age-specific data-point, containing leave-one-out cross validation statistics representing how well the model would predict that observation if it were left out of the fit.

These are computed with the loo package.

loo_indiv acts on the objects that are returned by running loo on disbayes objects. looi_disbayes acts directly on disbayes objects. Both of those functions return a data frame with LOO contributions for each data point.

Functions

looi_disbayes: Observation-level leave-one-out cross validation statistics for a disbayes model