Form a tidy data frame from the estimates from a disbayes fit

Simply call this after fitting disbayes, as, e.g.

res <- disbayes(...)
tidy(res)

# S3 method for disbayes
tidy(x, startyear = 1, ...)

# S3 method for disbayes_hier
tidy(x, ...)

Arguments

x
Object returned by disbayes

startyear
Only used for models with time trends. Numeric year represented by year 1 in the data. For example, set this to 1918 to convert years 1-100 to years 1918-2017.

...
Other arguments (currently unused)

Value

A data frame with one row per model parameter, giving summary statistics for the posterior distribution for that parameter. For array parameters, e.g. those that depend on age or area, then the age and area are returned in separate columns, to make it easier to summarise and plot the results, e.g. using ggplot2.

Model parameters might include, depending on the model specification,

cf Case fatality rates

inc Incidence rates

rem Remission rates

inc_par, cf_par, rem_par. Same as the above - should

inc_prob, rem_prob. Annual incidence and remission probabilities.

mort Annual mortality probability.

prev Prevalence probability.

state_probs State occupancy probabilities.

beta, beta_inc Coefficients of the spline basis for case fatality and incidence respectively.

lambda_cf, lambda_inc Smoothness parameters of the spline functions.

prevzero Prevalence at age zero

cfbase Case fatality at the baseline age (only in models where case fatality is increasing).

dcf Annual increments in case fatality (only in models where case fatality is increasing).

bias_loghr Log hazard ratio describing bias in case fatality between datasets (only in models where bias_model has been set).

For models with time trends:

cf_yr, inc_yr, state_probs_yr Case fatality rates, incidence rates and state occupancy probabilities in years prior to the current year. cf and inc refer to the rates for the current year, the one represented in the data.

Only for hierarchical models:

mean_inter, mean_slope,sd_inter,sd_slope. Mean and standard deviation for random effects distribution for the intercept and slope of log case fatality.

lambda_cf_male, lambda_inc_male. Smoothness of the additive gender effect on case fatality and incidence.

bareat Area-level contribution to spline basis coefficients.

barea Normalised spline basis coefficients.

Functions

tidy.disbayes_hier: Tidy method for hierarchical disbayes models